Rename previously unknown conversation fields

+++ WowPacketParser/Store/Objects/UpdateFields/LegacyImplementation/ConversationData.cs
@@ -65,13 +65,13 @@ namespace WowPacketParser.Store.Objects.UpdateFields.LegacyImplementation
                     {
                         if (ClientVersion.AddedInVersion(ClientVersionBuild.V8_0_1_27101))
                         {
-                            actor.Id = actors[i + 0];
+                            actor.Id = (int)actors[i + 0];
                             actor.CreatureID = actors[i + 0];
                             actor.CreatureDisplayInfoID = actors[i + 1];
                         }
                         else
                         {
-                            actor.Id = actors[i + 0];
+                            actor.Id = (int)actors[i + 0];
                             actor.CreatureID = actors[i + 1];
                             actor.CreatureDisplayInfoID = actors[i + 2];
                         }
@@ -95,7 +95,7 @@ namespace WowPacketParser.Store.Objects.UpdateFields.LegacyImplementation
 
         public class Actor : IConversationActor
         {
-            public uint Id { get; set; }
+            public int Id { get; set; }
             public uint CreatureID { get; set; }
             public uint CreatureDisplayInfoID { get; set; }
             public WowGuid ActorGUID { get; set; }
